requirements.txt
django
pyyaml
channels
channels_redis
django-crispy-forms
